Move details of draggable menu to DroppableIngredient
