+++ recipe-front-end/src/components/menu-planner/DayDetails.tsx
@@ -1,13 +1,9 @@
-import { Heading, Tooltip } from "@chakra-ui/react";
-import { faCheckSquare, faTimes, faTrash } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { Heading } from "@chakra-ui/react";
 import React from "react";
 import { connect } from 'react-redux';
-import { Dispatch } from "redux";
 import { Localisation } from "../../localisation/AppTexts";
-import { removeMenu, toggleMenuIngredientsBought, toggleMenuIngredientsBoughtReturn } from "../../redux/Actions";
-import { DayMenu, ReduxModel, RemoveMenuAction, ToggleMenuIngredientsBoughtAction } from "../../redux/Store";
-import { DroppableMenuItem } from "./DroppableIngredient";
+import { DayMenu, ReduxModel } from "../../redux/Store";
+import DroppableMenuItem from "./DroppableIngredient";
 import { filterForDate } from "./MenuPlanner";
 
 interface OwnProps {
@@ -19,11 +15,6 @@ interface ReduxProps {
     menuOfTheDay: DayMenu[];
 }
 
-interface ReduxActions {
-    removeMenu: (menu: DayMenu) => void;
-    toggleMenuIngredientsBought: toggleMenuIngredientsBoughtReturn;
-}
-
 function mapStateToProps(reduxStore: ReduxModel, ownProps: OwnProps): ReduxProps {
     const date = ownProps.date;
     return {
@@ -31,39 +22,18 @@ function mapStateToProps(reduxStore: ReduxModel, ownProps: OwnProps): ReduxProps
     }
 }
 
-function mapDispatchToProps(dispatch: Dispatch<RemoveMenuAction | ToggleMenuIngredientsBoughtAction>): ReduxActions {
-    return {
-        removeMenu: removeMenu(dispatch),
-        toggleMenuIngredientsBought: toggleMenuIngredientsBought(dispatch)
-    }
-}
-
-type Props = OwnProps & ReduxProps & ReduxActions;
+type Props = OwnProps & ReduxProps;
 
 function DayDetails(props: Props) {
     return <div className="menu-selected-day">
         <Heading as="h2">{Localisation.MENU_OF_THE_DAY}</Heading>
         {props.menuOfTheDay.length === 0 && <div>{Localisation.NOTHING_PLANNED_TODAY}</div>}
-        {props.menuOfTheDay.map((data, index) => {
-            return <div key={index}><DroppableMenuItem menu={data}>
-                <Tooltip label={Localisation.REMOVE}>
-                    <a href="#" onClick={() => props.removeMenu(data)}>
-                        <FontAwesomeIcon icon={faTrash} />
-                    </a>
-                </Tooltip>
-                <Tooltip label={data.ingredientsBought ?
-                    Localisation.YOU_HAVE_ALL_INGREDIENTS :
-                    Localisation.YOU_DONT_HAVE_ALL_INGREDIENTS
-                }>
-                    <a href="#" onClick={() => props.toggleMenuIngredientsBought([data], !data.ingredientsBought)}>
-                        <FontAwesomeIcon icon={data.ingredientsBought ? faCheckSquare : faTimes} />
-                    </a>
-                </Tooltip>
-            </DroppableMenuItem>
+        {props.menuOfTheDay.map((menu, index) => {
+            return <div key={index}><DroppableMenuItem menu={menu}/>
             </div>
         })}
         {props.children}
     </div>
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(DayDetails);
+export default connect(mapStateToProps)(DayDetails);
